Separate overrides from utils in array input tests
Context is moved to its own module to prevent circular import.
