+++ tests/array_inputs/README.md
@@ -3,3 +3,5 @@
 This directory contains the tests to check any functions / methods that accept array
 inputs. This is to ensure that the outputs / attributes are unchanged when broadcastable
 arrays are used in place of the full size arrays or when xarray DataArrays are used.
+
+In order to resolve failing tests refer to the `overrides.py` file.
