+++ tests/array_inputs/__init__.py
@@ -1,4 +1,4 @@
 """The module for testing array inputs (broadcasting, xarray, etc.).
 
-This file is required to make a module.
+The file overrides.py is used to define manual definitions to resolve failing tests.
 """
